+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 v0.6.2
 ------
 
+- [x] Fix: Issue #1 in Github
 - [x] Added key-binding(`u`) for reloading `music_dir`
 - [x] Fix: All option are now aligned properly in preferences view
 - [x] Added validators for preferences(Error message is shown if an error occurs)